Reserve timer channels used by TF-M
There are two timers Timer0 and Timer1 available in PSoC64 and the Timer0
contains 8 channel and Timer1 contains 24 Channels. TF-M regression tests
make use of Timer0 Channel 1 and Timer0 Channel 2. Therefore we are reserving
the timer channels used by TF-M. This is a temporary solution till we have a
way to allocate dedicated timers for TF-M and Mbed OS.

Signed-off-by: Devaraj Ranganna <devaraj.ranganna@arm.com>